Compendia and smasher tests are passing, gotta make sure qn_reference is still good.
