+++ workers/data_refinery_workers/processors/smasher.py
@@ -119,367 +119,6 @@ def _prepare_files(job_context: Dict) -> Dict:
     return job_context
 
 
-def _add_annotation_column(annotation_columns, column_name):
-    """Add annotation column names in place.
-    Any column_name that starts with "refinebio_" will be skipped.
-    """
-
-    if not column_name.startswith("refinebio_"):
-        annotation_columns.add(column_name)
-
-
-def _get_tsv_columns(job_context, samples_metadata):
-    """Returns an array of strings that will be written as a TSV file's
-    header. The columns are based on fields found in samples_metadata.
-
-    Some nested annotation fields are taken out as separate columns
-    because they are more important than the others.
-    """
-    tsv_start = log_state("start get tsv columns", job_context["job"])
-    refinebio_columns = set()
-    annotation_columns = set()
-    for sample_metadata in samples_metadata.values():
-        for meta_key, meta_value in sample_metadata.items():
-            if meta_key != 'refinebio_annotations':
-                refinebio_columns.add(meta_key)
-                continue
-
-            # Decompose sample_metadata["annotations"], which is an array of annotations!
-            for annotation in meta_value:
-                for annotation_key, annotation_value in annotation.items():
-                    # For ArrayExpress samples, take out the fields
-                    # nested in "characteristic" as separate columns.
-                    if (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                        and annotation_key == "characteristic"):
-                        for pair_dict in annotation_value:
-                            if 'category' in pair_dict and 'value' in pair_dict:
-                                _add_annotation_column(annotation_columns, pair_dict['category'])
-                    # For ArrayExpress samples, also take out the fields
-                    # nested in "variable" as separate columns.
-                    elif (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                          and annotation_key == "variable"):
-                        for pair_dict in annotation_value:
-                            if 'name' in pair_dict and 'value' in pair_dict:
-                                _add_annotation_column(annotation_columns, pair_dict['name'])
-                    # For ArrayExpress samples, skip "source" field
-                    elif (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                          and annotation_key == "source"):
-                        continue
-                    # For GEO samples, take out the fields nested in
-                    # "characteristics_ch1" as separate columns.
-                    elif (sample_metadata.get('refinebio_source_database', '') == "GEO"
-                          and annotation_key == "characteristics_ch1"): # array of strings
-                        for pair_str in annotation_value:
-                            if ':' in pair_str:
-                                tokens = pair_str.split(':', 1)
-                                _add_annotation_column(annotation_columns, tokens[0])
-                    # Saves all other annotation fields in separate columns
-                    else:
-                        _add_annotation_column(annotation_columns, annotation_key)
-
-    # Return sorted columns, in which "refinebio_accession_code" and "experiment_accession" are
-    # always first, followed by the other refinebio columns (in alphabetic order), and
-    # annotation columns (in alphabetic order) at the end.
-    refinebio_columns.discard('refinebio_accession_code')
-    log_state("end get tsv columns", job_context["job"], tsv_start)
-    return ['refinebio_accession_code', 'experiment_accession'] + sorted(refinebio_columns) \
-        + sorted(annotation_columns)
-
-
-def _add_annotation_value(row_data, col_name, col_value, sample_accession_code):
-    """Adds a new `col_name` key whose value is `col_value` to row_data.
-    If col_name already exists in row_data with different value, print
-    out a warning message.
-    """
-    # Generate a warning message if annotation field name starts with
-    # "refinebio_".  This should rarely (if ever) happen.
-    if col_name.startswith("refinebio_"):
-        logger.warning(
-            "Annotation value skipped",
-            annotation_field=col_name,
-            annotation_value=col_value,
-            sample_accession_code=sample_accession_code
-        )
-    elif col_name not in row_data:
-        row_data[col_name] = col_value
-    # Generate a warning message in case of conflicts of annotation values.
-    # (Requested by Dr. Jackie Taroni)
-    elif row_data[col_name] != col_value:
-        logger.warning(
-            "Conflict of values found in column %s: %s vs. %s" % (
-                col_name, row_data[col_name], col_value),
-            sample_accession_code=sample_accession_code
-        )
-
-
-def _get_experiment_accession(sample_accession_code, dataset_data):
-    for experiment_accession, samples in dataset_data.items():
-        if sample_accession_code in samples:
-            return experiment_accession
-    return ""  # Should never happen, because the sample is by definition in the dataset
-
-
-def _get_tsv_row_data(sample_metadata, dataset_data):
-    """Returns field values based on input sample_metadata.
-
-    Some annotation fields are treated specially because they are more
-    important.  See `_get_tsv_columns` function above for details.
-    """
-
-    sample_accession_code = sample_metadata.get('refinebio_accession_code', '')
-    row_data = dict()
-    for meta_key, meta_value in sample_metadata.items():
-        # If the field is a refinebio-specific field, simply copy it.
-        if meta_key != 'refinebio_annotations':
-            row_data[meta_key] = meta_value
-            continue
-
-        # Decompose sample_metadata["refinebio_annotations"], which is
-        # an array of annotations.
-        for annotation in meta_value:
-            for annotation_key, annotation_value in annotation.items():
-                # "characteristic" in ArrayExpress annotation
-                if (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                    and annotation_key == "characteristic"):
-                    for pair_dict in annotation_value:
-                        if 'category' in pair_dict and 'value' in pair_dict:
-                            col_name, col_value = pair_dict['category'], pair_dict['value']
-                            _add_annotation_value(row_data, col_name, col_value,
-                                                  sample_accession_code)
-                # "variable" in ArrayExpress annotation
-                elif (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                      and annotation_key == "variable"):
-                    for pair_dict in annotation_value:
-                        if 'name' in pair_dict and 'value' in pair_dict:
-                            col_name, col_value = pair_dict['name'], pair_dict['value']
-                            _add_annotation_value(row_data, col_name, col_value,
-                                                  sample_accession_code)
-                 # Skip "source" field ArrayExpress sample's annotation
-                elif (sample_metadata.get('refinebio_source_database', '') == "ARRAY_EXPRESS"
-                      and annotation_key == "source"):
-                    continue
-                # "characteristics_ch1" in GEO annotation
-                elif (sample_metadata.get('refinebio_source_database', '') == "GEO"
-                      and annotation_key == "characteristics_ch1"): # array of strings
-                    for pair_str in annotation_value:
-                        if ':' in pair_str:
-                            col_name, col_value = pair_str.split(':', 1)
-                            col_value = col_value.strip()
-                            _add_annotation_value(row_data, col_name, col_value,
-                                                  sample_accession_code)
-                # If annotation_value includes only a 'name' key, extract its value directly:
-                elif (isinstance(annotation_value, dict)
-                      and len(annotation_value) == 1 and 'name' in annotation_value):
-                    _add_annotation_value(row_data, annotation_key, annotation_value['name'],
-                                          sample_accession_code)
-                # If annotation_value is a single-element array, extract the element directly:
-                elif isinstance(annotation_value, list) and len(annotation_value) == 1:
-                    _add_annotation_value(row_data, annotation_key, annotation_value[0],
-                                          sample_accession_code)
-                # Otherwise save all annotation fields in separate columns
-                else:
-                    _add_annotation_value(row_data, annotation_key, annotation_value,
-                                          sample_accession_code)
-
-    row_data["experiment_accession"] = _get_experiment_accession(sample_accession_code,
-                                                                 dataset_data)
-
-    return row_data
-
-
-def _write_tsv_json(job_context, metadata):
-    """Writes tsv files on disk.
-    If the dataset is aggregated by species, also write species-level
-    JSON file.
-    """
-
-    # Uniform TSV header per dataset
-    columns = _get_tsv_columns(job_context, metadata['samples'])
-
-    # Per-Experiment Metadata
-    if job_context["dataset"].aggregate_by == "EXPERIMENT":
-        tsv_paths = []
-        for experiment_title, experiment_data in metadata['experiments'].items():
-            experiment_dir = job_context["output_dir"] + experiment_title + '/'
-            experiment_dir = experiment_dir.encode('ascii', 'ignore')
-            os.makedirs(experiment_dir, exist_ok=True)
-            tsv_path = experiment_dir.decode("utf-8") + 'metadata_' + experiment_title + '.tsv'
-            tsv_path = tsv_path.encode('ascii', 'ignore')
-            tsv_paths.append(tsv_path)
-            with open(tsv_path, 'w', encoding='utf-8') as tsv_file:
-                dw = csv.DictWriter(tsv_file, columns, delimiter='\t')
-                dw.writeheader()
-                for sample_accession_code, sample_metadata in metadata['samples'].items():
-                    if sample_accession_code in experiment_data['sample_accession_codes']:
-                        row_data = _get_tsv_row_data(sample_metadata, job_context["dataset"].data)
-                        dw.writerow(row_data)
-        return tsv_paths
-    # Per-Species Metadata
-    elif job_context["dataset"].aggregate_by == "SPECIES":
-        tsv_paths = []
-        for species in job_context['input_files'].keys():
-            species_dir = job_context["output_dir"] + species + '/'
-            os.makedirs(species_dir, exist_ok=True)
-            samples_in_species = []
-            tsv_path = species_dir + "metadata_" + species + '.tsv'
-            tsv_paths.append(tsv_path)
-            with open(tsv_path, 'w', encoding='utf-8') as tsv_file:
-                dw = csv.DictWriter(tsv_file, columns, delimiter='\t')
-                dw.writeheader()
-                for sample_metadata in metadata['samples'].values():
-                    if sample_metadata.get('refinebio_organism', '') == species:
-                        row_data = _get_tsv_row_data(sample_metadata, job_context["dataset"].data)
-                        dw.writerow(row_data)
-                        samples_in_species.append(sample_metadata)
-
-            # Writes a json file for current species:
-            if len(samples_in_species):
-                species_metadata = {
-                    'species': species,
-                    'samples': samples_in_species
-                }
-                json_path = species_dir + "metadata_" + species + '.json'
-                with open(json_path, 'w', encoding='utf-8') as json_file:
-                    json.dump(species_metadata, json_file, indent=4, sort_keys=True)
-        return tsv_paths
-    # All Metadata
-    else:
-        all_dir = job_context["output_dir"] + "ALL/"
-        os.makedirs(all_dir, exist_ok=True)
-        tsv_path = all_dir + 'metadata_ALL.tsv'
-        with open(tsv_path, 'w', encoding='utf-8') as tsv_file:
-            dw = csv.DictWriter(tsv_file, columns, delimiter='\t')
-            dw.writeheader()
-            for sample_metadata in metadata['samples'].values():
-                row_data = _get_tsv_row_data(sample_metadata, job_context["dataset"].data)
-                dw.writerow(row_data)
-        return [tsv_path]
-
-
-def _quantile_normalize(job_context: Dict, ks_check=True, ks_stat=0.001) -> Dict:
-    """
-    Apply quantile normalization.
-
-    """
-    # Prepare our QN target file
-    organism = job_context['organism']
-    qn_target = utils.get_most_recent_qn_target_for_organism(organism)
-
-    if not qn_target:
-        logger.error("Could not find QN target for Organism!",
-            organism=organism,
-            dataset_id=job_context['dataset'].id,
-            processor_job_id=job_context["job"].id,
-        )
-        job_context['dataset'].success = False
-        job_context['job'].failure_reason = "Could not find QN target for Organism: " + str(organism)
-        job_context['dataset'].failure_reason = "Could not find QN target for Organism: " + str(organism)
-        job_context['dataset'].save()
-        job_context['job'].success = False
-        job_context['failure_reason'] = "Could not find QN target for Organism: " + str(organism)
-        return job_context
-    else:
-        qn_target_path = qn_target.sync_from_s3()
-        qn_target_frame = pd.read_csv(qn_target_path, sep='\t', header=None,
-                                      index_col=None, error_bad_lines=False)
-
-        # Prepare our RPy2 bridge
-        pandas2ri.activate()
-        preprocessCore = importr('preprocessCore')
-        as_numeric = rlang("as.numeric")
-        data_matrix = rlang('data.matrix')
-
-        # Convert the smashed frames to an R numeric Matrix
-        # and the target Dataframe into an R numeric Vector
-        target_vector = as_numeric(qn_target_frame[0])
-        merged_matrix = data_matrix(job_context['merged_no_qn'])
-
-        # Perform the Actual QN
-        reso = preprocessCore.normalize_quantiles_use_target(
-                                            x=merged_matrix,
-                                            target=target_vector,
-                                            copy=True
-                                        )
-
-        # Verify this QN, related: https://github.com/AlexsLemonade/refinebio/issues/599#issuecomment-422132009
-        set_seed = rlang("set.seed")
-        combn = rlang("combn")
-        ncol = rlang("ncol")
-        ks_test = rlang("ks.test")
-        which = rlang("which")
-
-        set_seed(123)
-
-        n = ncol(reso)[0]
-        m = 2
-        if n >= m:
-            combos = combn(ncol(reso), 2)
-
-            # Convert to NP, Shuffle, Return to R
-            ar = np.array(combos)
-            np.random.shuffle(np.transpose(ar))
-            nr, nc = ar.shape
-            combos = ro.r.matrix(ar, nrow=nr, ncol=nc)
-
-            # adapted from
-            # https://stackoverflow.com/questions/9661469/r-t-test-over-all-columns
-            # apply KS test to randomly selected pairs of columns (samples)
-            for i in range(1, min(ncol(combos)[0], 100)):
-                value1 = combos.rx(1, i)[0]
-                value2 = combos.rx(2, i)[0]
-
-                test_a = reso.rx(True, value1)
-                test_b = reso.rx(True, value2)
-
-                # RNA-seq has a lot of zeroes in it, which
-                # breaks the ks_test. Therefore we want to
-                # filter them out. To do this we drop the
-                # lowest half of the values. If there's
-                # still zeroes in there, then that's
-                # probably too many zeroes so it's okay to
-                # fail.
-                median_a = np.median(test_a)
-                median_b = np.median(test_b)
-
-                # `which` returns indices which are
-                # 1-indexed. Python accesses lists with
-                # zero-indexes, even if that list is
-                # actually an R vector. Therefore subtract
-                # 1 to account for the difference.
-                test_a = [test_a[i-1] for i in which(test_a > median_a)]
-                test_b = [test_b[i-1] for i in which(test_b > median_b)]
-
-                # The python list comprehension gives us a
-                # python list, but ks_test wants an R
-                # vector so let's go back.
-                test_a = as_numeric(test_a)
-                test_b = as_numeric(test_b)
-
-                ks_res = ks_test(test_a, test_b)
-                statistic = ks_res.rx('statistic')[0][0]
-                pvalue = ks_res.rx('p.value')[0][0]
-
-                job_context['ks_statistic'] = statistic
-                job_context['ks_pvalue'] = pvalue
-
-                # We're unsure of how strigent to be about
-                # the pvalue just yet, so we're extra lax
-                # rather than failing tons of tests. This may need tuning.
-                if ks_check:
-                    if statistic > ks_stat or pvalue < 0.8:
-                        job_context['ks_warning'] = ("Failed Kolmogorov Smirnov test! Stat: " +
-                                        str(statistic) + ", PVal: " + str(pvalue))
-        else:
-            logger.warning("Not enough columns to perform KS test - either bad smash or single saple smash.",
-                dataset_id=job_context['dataset'].id)
-
-        # And finally convert back to Pandas
-        ar = np.array(reso)
-        new_merged = pd.DataFrame(ar, columns=job_context['merged_no_qn'].columns, index=job_context['merged_no_qn'].index)
-        job_context['merged_qn'] = new_merged
-        merged = new_merged
-    return job_context
-
 def sync_quant_files(output_path, files_sample_tuple, job_context: Dict):
     """ Takes a list of ComputedFiles and copies the ones that are quant files to the provided directory.
         Returns the total number of samples that were included """
@@ -572,7 +211,7 @@ def _inner_join(job_context: Dict) -> pd.DataFrame:
 # Nooooo, I think I wanna call process frame from compendia, not this.
 # Therefore this can just always do an inner join and then the
 # compendia can do its own joining.
-def smash_key(job_context: Dict,
+def _smash_key(job_context: Dict,
               key: str,
               input_files: List[ComputedFile],
               how="inner") -> Dict:
@@ -584,6 +223,8 @@ def smash_key(job_context: Dict,
         Scale features with sci-kit learn
         Transpose again such that samples are columns and genes are rows
     """
+    start_smash = log_state("end build all frames", job_context["job"])
+
     # Check if we need to copy the quant.sf files
     if job_context['dataset'].quant_sf_only:
         outfile_dir = job_context["output_dir"] + key + "/"
@@ -593,15 +234,6 @@ def smash_key(job_context: Dict,
         return job_context
 
     job_context = smashing_utils.process_frames_for_key(key, input_files, job_context)
-    if len(job_context['all_frames']) < 1:
-        logger.error("Was told to smash a key with no frames!",
-                       job_id=job_context['job'].id,
-                       key=key)
-        # TODO: is this the proper way to handle this? I can see us
-        # not wanting to fail an entire dataset because one experiment
-        # had a problem, but I also think it could be problematic to
-        # just skip an experiment and pretend nothing went wrong.
-        return job_context
 
     # Combine the two technologies into a single list of dataframes.
     ## Extend one list rather than adding the two together so we don't
@@ -611,6 +243,16 @@ def smash_key(job_context: Dict,
     job_context.pop('microarray_frames')
     ## Change the key of the now-extended list
     job_context['all_frames'] = job_context.pop('rnaseq_frames')
+
+    if len(job_context['all_frames']) < 1:
+        logger.error("Was told to smash a key with no frames!",
+                       job_id=job_context['job'].id,
+                       key=key)
+        # TODO: is this the proper way to handle this? I can see us
+        # not wanting to fail an entire dataset because one experiment
+        # had a problem, but I also think it could be problematic to
+        # just skip an experiment and pretend nothing went wrong.
+        return job_context
 
     if how == "inner":
         merged = _inner_join(job_context)
@@ -631,7 +273,7 @@ def smash_key(job_context: Dict,
         try:
             job_context['merged_no_qn'] = merged
             job_context['organism'] = job_context['dataset'].get_samples().first().organism
-            job_context = _quantile_normalize(job_context)
+            job_context = smashing_utils.quantile_normalize(job_context)
             merged = job_context.get('merged_qn', None)
 
             # We probably don't have an QN target or there is another error,
@@ -732,7 +374,7 @@ def _smash_all(job_context: Dict, how="inner") -> Dict:
 
         # Write samples metadata to TSV
         try:
-            tsv_paths = _write_tsv_json(job_context, metadata, job_context["output_dir"])
+            tsv_paths = smashing_utils.write_tsv_json(job_context, metadata)
             job_context['metadata_tsv_paths'] = tsv_paths
             # Metadata to JSON
             metadata['created_at'] = datetime.utcnow().strftime('%Y-%m-%dT%H:%M:%S')
@@ -973,7 +615,7 @@ def smash(job_id: int, upload=True) -> None:
                             },
                        [utils.start_job,
                         smashing_utils.prepare_files,
-                        _smash,
+                        _smash_all,
                         _upload,
                         _notify,
                         _update_result_objects,
